feat(wiz): batch worksheet-table creation in one /ws/table request
Change POST /api/wiz/ws/table to accept a batch of tables: WorksheetTableRequest becomes { worksheetId, List<WorksheetTable> tables } and the tables are created against one worksheet loaded/persisted once. Each table is built independently — on failure its assembly is rolled back and it is reported success:false; a table whose dependency failed or is absent is skipped with a clear message; only succeeded tables are persisted. Adds WorksheetTable, WorksheetTablesResponse, and WorksheetTableServiceBatchTest.
